set up testing infrastructure for remote store
