+++ pyproject.toml
@@ -112,8 +112,11 @@ extra-dependencies = [
     "pytest-cov",
     "msgpack",
     "lmdb",
+    "s3fs",
     "pytest-asyncio",
-    "moto",
+    "moto[s3]",
+    "flask-cors",
+    "flask",
     "requests",
     "mypy"
 ]
